Upgrade Elasticsearch to version 5
There are a few reasons why this seems more complicated than it should
be. Elasticsearch 5 needs a higher `vm.max_map_count` on the host than
is set by default. A privileged init container is used to set this. The
official Elasticsearch 5 image has x-pack enabled by default. This can
be disabled by passing environment variables, but the environment
variable names have a period in them and kubernetes does not allow this.
A custom Elasticsearch image is used for the sole reason of disabling
x-pack.